requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
allure-pytest==2.15.0
pytest-xdist==3.5.0